Move processLiquidStrings to liquid-render.js and rename filter
- Move processLiquidStrings to liquid-render.js, reusing existing Liquid
  instance instead of creating a duplicate
- Rename filter from snippet_blocks_with_context to snippet_blocks
- Trim verbose comment in snippet-block.html
