+++ src/_includes/design-system/snippet-block.html
@@ -2,11 +2,10 @@
 Renders blocks from a named snippet file. Used by render-block.html.
 The snippet block is transparent — blocks.html skips the outer <section>
 wrapper for this type, so each inner block renders its own section directly.
-Liquid expressions in block values (e.g. {{ title }}) are resolved against
-the current page context before rendering.
+Liquid expressions in block values are resolved against the page context.
 {%- endcomment -%}
 
-{%- assign processed_blocks = block.reference | snippet_blocks_with_context -%}
+{%- assign processed_blocks = block.reference | snippet_blocks -%}
 {%- if processed_blocks.size > 0 -%}
   {%- include "design-system/blocks.html", blocks: processed_blocks -%}
 {%- endif -%}
